requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
openpyxl>=3.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
streamlit>=1.30.0
